Enable aborting on error by returning the status from surroundings#Lines#SurroundCommand() and using ingo/err.vim
CHG: Also allow transforming after the before- and after-lines have been added. Restructure a:Transformer argument into generic a:options, with keys for the old a:options.TransformerBefore and the new a:options.TransformerAfter.
ENH: The Transformer(s) can now change the amount of lines; the algorithm now deals with that by calculating the offset.