ingo#query#get#{Register,Mark}(): Avoid throwing E523 on invalid user input
The functions may be called from a :map-expression, which disallows :normal. But that is used to generate the beep if the user entered an invalid register / mark. Instead of propagating the exception, catch it and convert to the proper a:errorRegister / empty string. The underlying ingo#query#get#Char() supports a:options.isBeepOnInvalid, but exposing that to these functions would be overkill (and require client changes / awareness of this issue).