ingo#cmdargs#command#Parse(): Improve parsing of command sequences
This somehow previously failed when there's whitespace between a previous command and the | delimiter:
:echo ingo#cmdargs#command#Parse('new|verbose s /foo/bar/g', '.*$') " OK
:echo ingo#cmdargs#command#Parse('new |verbose s /foo/bar/g', '.*$') " FAIL: Parses "new" as command.

The cause is that /^\s*/ yields an empty match instead of /\\\@<!|\s*/ matching. Break the parsing into two iterations, and try with the |-separation first. This should also make troubleshooting easier, as we now use more intermediate variables.