ENH: ingo#plugin#rendered#List(): Allow dynamic a:renderer.options; add strict / short options for BraceExpressionRenderer
Turn a:renderer.options into a method a:renderer.options() that is re-invoked on each iteration and therefore can change based on chosen commands.

Add commands to BraceExpressionRenderer that enable strict or short rendering mode. Some clients have dedicated mappings for both variants, but not all, and it can be convenient to switch interactively when one realizes that the current mode doesn't yield good results.