Tests: Allow to choose compatibility emulations for certain functions
This is mostly for unit tests that run in (usually recent) Vim versions that don't require the compatibility functions, but instead just forward the call to the built-in function. Testing this is pointless; we need to compare the emulation's behavior with the built-in.
Consider comma-separated list of function names in g:IngoLibrary_CompatFor. Read those into a Dictionary, and skip the forwarding wrapper for those function names; using the compatibility function instead.