ingo#plugin#rendered#Subset#Filter(): Properly unescape separator character in /pattern/
To be consistent with the built-in commands like :substitute. It's unlikely that a user uses the same delimiter (e.g. /foo\/bar/) when the delimiter can be freely chosen, and for many delimiters, the backslash in front of it is not critical (but it is for e.g. +.), but better be correct.