ingo#regexp#length#Project(): FIX: Combining of multi with previous pattern was wrong, change algorithm
The multi must apply to (min/max of) the preceding literal character or group. The current implementation applied it to the previous literal text, as if there were an invisible group around it.
In s:ProjectBranch(), start with the split into prefix/groups/suffix now, and apply a multi following a group.
Multis of literals are now handled in s:ProjectUngroupedPattern(); the original code from s:ProjectBranch() moved here. Also need to add up the min/max here instead of determining the overall from concatenated multis; this was wrong. s:ProjectMultis() does not need to multiply (that's only for groups, in s:Multiply()); instead, it adds the literal min/max plus the multi, minus the one for the multiplied atom itself.