Tests: Add coverage of adding patterns considering min-max lengths
The ingo#regexp#length#Project() tests check the determination of the correct min-max values for multis, but ingo#regexp#split#AddPatternByProjectedMatchLength() does its own comparison of max (or min, if max is endless) in branches against min of the addition.
The added tests cover that, and show the somewhat unexpected behavior of \{M,N} multis. I still don't wanna introduce more complexity for them, as they are rare, the current behavior isn't completely stupid, and all of this is an approximation, anyway.