Reimplementation: ingo#strdisplaywidth#strleft()
Avoid the downstream check and truncation by substitute()'ing any exceeding characters away. In order to exclude a character that does not entirely fit the passed width, we match for larger than the passed width plus one, and the character that comes before, similar to what we do in ingo#strdisplaywidth#HasMoreThan().