FIX: Make ingo#cmdline#showmode#TemporaryNoShowMode() work again
The CursorMovedI event is triggered immediately after the abbreviation is expanded (in my IHint.vim plugin), and 'showmode' is directly restored. I dunno how this worked before (but I remember it did).
Exclude the current cursor position by setting and comparing a record. A difficulty here is that we need the record _after_ the typed character, so one to the right. Apparently, we cannot grab that character, because it's not there yet. Fortunately, abbreviation expansion is mostly triggered by whitespace and ASCII characters, so just add one to the column and it should be good in most cases.