ENH: ingo#cmdrangeconverter#LineToBufferRange(): Also allow Funcref argument
Functions naturally apply to a single line, so allowing a range conversion makes sense here as well, and I actually needed this for composing the range conversion with register manipluation in AdvancedDeleteMappings.vim.
We don't need to extend the other (ingo#cmdrangeconverter#BufferToLineRange()) function, as there are no functions that default to a buffer range.