ingo#text#frompattern#Get(): Inline s:UniqueAdd(), add s:PredicateCheck(), and only invoke the latter if non-unique match
This avoids predicate invocations that later would be discarded due to duplicate matches, which makes for better semantics if the predicate maintains some state across invocations (e.g. counting the number of matches).

Also, there was a small bug lurking in the previous implementation with a:isOnlyFirstMatch was set: a non-unique match would still discard the remainder of the line (and thereby potentially discarding following non-unique matches) because s:UniqueAdd() did not indicate the uniqueness and the code always proceeded.