BUG: ingo#join#Lines() inserts the separator before the line's last character when joining with a following empty line and  a:isKeepSpace
When the following line is empty or just consists of whitespace, the cursor is at the end of the current line after joining, and "a" instead of "i" has to be used to insert the separator.
Note: We cannot use ingo#cursor#IsAtEndOfLine() here, because when joining with a single character line (for this special case: a line containing just a ")"), the cursor also would be a the end, but "i" needs to be used.