ingo#subs#apply#FlexibleExpression(): Allow filtering and adding of processed matches / splits by having expression return a List
As the algorithm works by splitting into items and separators and then map'ping one through the expression, we have to find a joiner for any added items. For ^{pattern}^, we can simply obtain the first match from the original text. For /{pattern}/, the first element of split() will be it (if such exists).