FIX: Directly echo previous errors instead of concatenating them
The problem is that :echoerr does not handle newlines (\n -> ^@, \r -> ^M), and as the "echoerr ingo#err#Get()" is used pervasively by clients, it's hard to turn than into a loop, also because of the exception semantics of :echoerr (so subsequent ones might be suppressed).
Instead, if a second (third, ...) error is encountered, :echomsg the previous one with error highlighting, and move the next error into the slot, so that only the last error is used with :echoerr.