ingo#file#GetLines(): Robustness: Empty a:filespec throws E17
The canonicalization transforms the empty filespec into the current working directory, and that one has a modification time, so a read is attempted, which then fails with "E17: file is a directory", which isn't handled by the :catch clause.
Check for empty a:filespec at the top and also handle the E17 exception.