Pass the maxItemNum to a Funcref for s:GetSource(s:historySources)
For a List, we need a maximum of 9, and for a direct recall, we just need the count'th one. I don't want to introduce two separate APIs for the list and the indexed get; I think a combination that just asks for a List slice is a good compromise. Clients can completely ignore this and return the full List. Other clients may have to iterate over a source and for those the opportunity to stop early when enough is produced is valuable, though.