BUG: g:IngoLibrary_QueryMotionCustomMotions appendagePattern is not considered
Being a custom motion, the maparg() branch already gets taken, and the fact that the elements have been added to s:builtInMotions (not very clean to start with) does not matter here.
We cannot move the built-in branch above the custom motion one, as custom motions do have precedence.
Rather, keep the g:IngoLibrary_QueryMotionCustomMotions contents out of s:builtInMotions, and duplicate the code for the appendage handling to that branch.