BUG: Shorter built-in motions eclipse longer custom motions
Add mapcheck() to avoid taking that branch, again considering exceptions in g:IngoLibrary_QueryMotionIgnoredMotions. Problem is that without a timeout, we cannot handle built-in mappings that are subsets of longer custom motions. However, being motions, such setup of requiring timeouts for navigation would be very cumbersome, and therefore hopefully rare.