ingo#range#Get(): Don't clobber the search history
For all existing clients, this would be an undesired side effect.
We cannot simply use the compat function, as the fallback implementation is not meant for use inside a function. Instead, copy the check from ingo#range#lines#Get() and remove the last search history item if the search pattern got changed.